Урок 6. Оставили в меню вызов фрагмента с описанием программы (заглушка). Создали файл сервиса для дальнейшей обработки сетевого запроса погоды через него.
