+++ app/src/main/java/by/ealipatov/kotlin/weatherfromealipatov/view/AboutFragment.kt
@@ -1,0 +1,38 @@
+package by.ealipatov.kotlin.weatherfromealipatov.view
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import by.ealipatov.kotlin.weatherfromealipatov.databinding.FragmentAboutBinding
+
+class AboutFragment: Fragment() {
+
+    private var _binding: FragmentAboutBinding? = null
+    private val binding: FragmentAboutBinding
+        get() {
+            return _binding!!
+        }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        _binding = null
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = FragmentAboutBinding.inflate(inflater)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        binding.aboutText.text = "Проект программы отображения погоды по городам"
+    }
+
+}
