+++ app/src/main/java/by/ealipatov/kotlin/weatherfromealipatov/view/LoadWeatherService.kt
@@ -1,0 +1,14 @@
+package by.ealipatov.kotlin.weatherfromealipatov.view
+
+import android.app.IntentService
+import android.content.Intent
+
+class LoadWeatherService : IntentService("LoadWeatherService") {
+
+    @Deprecated("Deprecated in Java")
+    override fun onHandleIntent(intent: Intent?) {
+
+    }
+
+
+}
