+++ app/src/main/res/menu/main_menu.xml
@@ -3,13 +3,9 @@
     xmlns:tools="http://schemas.android.com/tools"
     tools:context=".MainActivity">
     <item
-        android:id="@+id/new_fragment"
+        android:id="@+id/about"
         android:orderInCategory="100"
-        android:title="@string/new_fragment"
+        android:title="@string/about"
         app:showAsAction="never"/>
 
-    <item
-        android:id="@+id/threads"
-        android:title="@string/threads"
-        app:showAsAction="never" />
 </menu>
