+++ app/src/main/res/values/strings.xml
@@ -36,4 +36,5 @@
     <string name="thunderstorm_with_rain">Дождь с грозой</string>
     <string name="thunderstorm_with_hail">Гроза с градом</string>
     <string name="new_fragment">Новый фрагмент</string>
+    <string name="about">О программе</string>
 </resources>
